+++ tests/e2e/support/api/graph/user.ts
@@ -1,8 +1,18 @@
-import { checkResponseStatus, checkOCJsonStatus, request } from '../http'
-import { User } from '../../types'
+import { checkResponseStatus, request } from '../http'
+import { Me, User } from '../../types'
 import join from 'join-path'
 
-export const createUser = async ({ user, admin }: { user: User; admin: User }): Promise<void> => {
+export const me = async ({ user }: { user: User }): Promise<Me> => {
+  const response = await request({
+    method: 'GET',
+    path: join('graph', 'v1.0', 'me'),
+    user
+  })
+
+  return await response.json()
+}
+
+export const createUser = async ({ user, admin }: { user: User; admin: User }): Promise<User> => {
   const body = JSON.stringify({
     displayName: user.displayName,
     mail: user.email,
@@ -13,10 +23,11 @@ export const createUser = async ({ user, admin }: { user: User; admin: User }): 
   const response = await request({
     method: 'POST',
     path: join('graph', 'v1.0', 'users'),
-    body: body,
+    body,
     user: admin
   })
   checkResponseStatus(response, 'Failed while creating user')
+  return user
 }
 
 export const deleteUser = async ({ user, admin }: { user: User; admin: User }): Promise<User> => {
